refactor(demos): migrate wave 6 (defaults factories) to the shared style system
Move defaults-sync-factory and defaults-async-factory flat to folder, both
landing with zero bespoke residue. Adds two shared fragments seeded here and
reused by upcoming waves (autosave, on-change, step-slots, validate-on-modes):

- header: card/section title row (title left, action or status pill right)
- defgrid: compact two-column key-value <dl>, pairing with deflist's stacked
  value boxes for the "reactive state" readout idiom

Extends existing fragments rather than re-encoding state styling per demo:
input/select gain :read-only and :disabled, card gains an [aria-busy] dim,
badge gains a .busy (warning) variant beside .pending.

Collapses the demos' synonyms onto canonical vocabulary: .layout to the demo
layout split modifier, the dark terminal <pre> to the light readout, .toggle
to row compact, .error-row to banner failed, the muted placeholder to hint,
and the async badges to badge busy/success/error.
